+++ src/main/java/com/smarthome/infrastructure/database/adapter/LocationJpaRepository.java
@@ -3,5 +3,5 @@ package com.smarthome.infrastructure.database.adapter;
 import com.smarthome.infrastructure.database.entities.location.LocationJpaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationJpaRepository extends JpaRepository<LocationJpaEntity, Integer> {
+public interface LocationJpaRepository extends JpaRepository<LocationJpaEntity, Long> {
 }
